Split CLI into a package with focused submodules
Replaced zavod/cli.py with a zavod/cli/ package split into:
- etl.py: crawl, validate, export, publish, run
- dedupe.py: xref, resolver-prune, dedupe, explode/merge/dedupe-edges
- util.py: load-db, dump-file, clear, summarize
- wd_up.py: wd-up

Shared helpers (DatasetPath, _load_dataset, etc.) live in __init__.py,
which imports the submodules at the bottom to register commands. No
functional changes.
